requirements.txt
python-dotenv~=1.0.1
psycopg~=3.1.19
fastapi~=0.111.0
pydantic~=2.7.4
jose~=1.0.0